refactor(work-orders): extract parseList to shared utils (dedup)
`parseList` lived in both WorkOrderDetail and PrintableWorkOrder as
identical copies (QA finding #4). Extracted into
`features/work-orders/utils.ts` so the fix for Array<object> inputs
stays in one place and the two consumers can't drift again.
